third_party/WebKit/LayoutTests/svg/dom/SVGViewSpec-expected.txt

CONSOLE ERROR: line 87: Uncaught SyntaxError: Unexpected end of input
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
